Rawr3.BossHandler: UI work for better Bindings interaction
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C51986

+++ Rawr.Base/Bosses/BossList.cs
@@ -412,7 +412,7 @@ namespace Rawr {
             retboss.Name = "The Easiest Boss";
             value = 0f; foreach (BossHandler boss in passedList) { value = Math.Max(value, boss.BerserkTimer); } retboss.BerserkTimer = (int)Math.Ceiling(value);
             value = passedList[0].Health; foreach (BossHandler boss in passedList) { value = Math.Min(value, boss.Health); } retboss.Health = value;
-            value = passedList[0].Armor; foreach (BossHandler boss in passedList) { value = Math.Min(value, boss.Armor); } retboss.Armor = value;
+            value = passedList[0].Armor; foreach (BossHandler boss in passedList) { value = Math.Min(value, boss.Armor); } retboss.Armor = (int)value;
             retboss.UseParryHaste = false;
             // Resistance
             foreach (ItemDamageType t in DamageTypes) {
@@ -555,7 +555,7 @@ namespace Rawr {
             retboss.Name = "The Average Boss";
             value = 0f; foreach (BossHandler boss in passedList) { value += boss.BerserkTimer; } value /= passedList.Length; retboss.BerserkTimer = (int)Math.Floor(value);
             value = 0f; foreach (BossHandler boss in passedList) { value += boss.Health; } value /= passedList.Length; retboss.Health = value;
-            value = 0f; foreach (BossHandler boss in passedList) { value += boss.Armor; } value /= passedList.Length; retboss.Armor = value;
+            value = 0f; foreach (BossHandler boss in passedList) { value += boss.Armor; } value /= passedList.Length; retboss.Armor = (int)value;
             use = false; foreach (BossHandler boss in passedList) { use |= boss.UseParryHaste; } retboss.UseParryHaste = use;
             // Resistance
             foreach (ItemDamageType t in DamageTypes) {
@@ -731,7 +731,7 @@ namespace Rawr {
             retboss.Name = "The Hardest Boss";
             value = passedList[0].BerserkTimer; foreach (BossHandler boss in passedList) { value = Math.Min(value, boss.BerserkTimer); } retboss.BerserkTimer = (int)Math.Floor(value);
             value = 0f; foreach (BossHandler boss in passedList) { value = Math.Max(value, boss.Health); } retboss.Health = value;
-            value = 0f; foreach (BossHandler boss in passedList) { value = Math.Max(value, boss.Armor); } retboss.Armor = value;
+            value = 0f; foreach (BossHandler boss in passedList) { value = Math.Max(value, boss.Armor); } retboss.Armor = (int)value;
             retboss.UseParryHaste = true;
             // Resistance
             foreach (ItemDamageType t in DamageTypes) {
